Try to get rid of the postcss error
Package postcss can't be external
The request postcss matches serverExternalPackages (or the default list).
The request could not be resolved by Node.js from the project directory.
Packages that should be external need to be installed in the project directory, so they can be resolved from the output files.
Try to install it into the project directory by running npm install postcss from the project directory.